feat(env): nouvel environnement formation

+++ _dist_restore-dist-dump-from-prod-transfer.sh
@@ -7,7 +7,7 @@ if [ -z "$DOMIFA_ENV_ID" ]; then
     echo "[ERROR] DOMIFA_ENV_ID is not defined in .env"
     exit 1
 fi
-if [ "$DOMIFA_ENV_ID" !== "preprod" && "$DOMIFA_ENV_ID" !== "formation" ]; then
+if [ "$DOMIFA_ENV_ID" != "preprod" && "$DOMIFA_ENV_ID" != "formation" ]; then
     echo "[ERROR] script not allowed in this environment (only for 'preprod' & 'formation')"
     exit 1
 else
@@ -49,7 +49,7 @@ echo ""
 echo "#############################################################################"
 echo "[INFO] Stop backend & frontend"
 echo ""
-(set -x && sudo docker stop master_backend_1 master_frontend_1)
+(set -x && sudo docker stop ${DOCKER_COMPOSE_PROJECT_NAME}_backend_1 ${DOCKER_COMPOSE_PROJECT_NAME}_frontend_1)
 
 echo ""
 echo "#############################################################################"
@@ -59,9 +59,19 @@ echo ""
 # copier le dump dans le container
 (set -x && sudo docker cp $DUMP_FILE_MONGO ${DOCKER_COMPOSE_PROJECT_NAME}_mongo_1:/tmp)
 
-# restaurer le dump
-sudo docker exec -t ${DOCKER_COMPOSE_PROJECT_NAME}_mongo_1 bash -c "set -x && mongorestore --nsInclude 'domifa.*' --nsFrom 'domifa.*' --nsTo 'domifa.*' --drop --gzip --archive=/tmp/domifa_PROD.mongo.gz"
+if [ $? -eq 1 ]; then
+    echo "[ERROR] exit"
+    exit 3
+fi
 
+# restaurer le dump
+sudo docker exec -t ${DOCKER_COMPOSE_PROJECT_NAME}_mongo_1 bash -c "set -x && mongorestore --nsInclude 'domifa.*' --nsFrom 'domifa.*' --nsTo 'domifa.*' -u \$MONGO_INITDB_ROOT_USERNAME -p \$MONGO_INITDB_ROOT_PASSWORD --authenticationDatabase=admin --drop --gzip --archive=/tmp/domifa_PROD.mongo.gz"
+
+if [ $? -eq 1 ]; then
+    echo "[ERROR] exit"
+    ex
+    it 3
+fi
 echo ""
 echo "#############################################################################"
 echo "[INFO] Restore POSTGRES dump from $DUMP_FILE_POSTGRES..."
@@ -69,8 +79,26 @@ echo ""
 
 # copier le dump dans le container
 (set -x && sudo docker cp $DUMP_FILE_POSTGRES ${DOCKER_COMPOSE_PROJECT_NAME}_postgres_1:/tmp)
+if [ $? -eq 1 ]; then
+    echo "[ERROR] exit"
+    exit 3
+fi
 
 # restaurer le dump
 sudo docker exec -t ${DOCKER_COMPOSE_PROJECT_NAME}_postgres_1 bash -c "set -x && psql --username ${POSTGRES_USERNAME} --dbname postgres -c \"DROP DATABASE IF EXISTS \${POSTGRES_DB}\""
+if [ $? -eq 1 ]; then
+    echo "[ERROR] exit"
+    exit 3
+fi
+
 sudo docker exec -t ${DOCKER_COMPOSE_PROJECT_NAME}_postgres_1 bash -c "set -x && psql --username ${POSTGRES_USERNAME} --dbname postgres -c \"CREATE DATABASE \${POSTGRES_DB}\""
+if [ $? -eq 1 ]; then
+    echo "[ERROR] exit"
+    exit 3
+fi
+
 sudo docker exec -t ${DOCKER_COMPOSE_PROJECT_NAME}_postgres_1 bash -c "set -x && pg_restore --username=${POSTGRES_USERNAME} --no-owner --role=${POSTGRES_USERNAME} --exit-on-error --verbose --dbname=\${POSTGRES_DB} /tmp/domifa_PROD.postgres.custom.gz"
+if [ $? -eq 1 ]; then
+    echo "[ERROR] exit"
+    exit 3
+fi
